move the authorization token to search page
